refactor(security-hub): dedupe SEVERITY_RANK + split SecuritySafesTable
- primitives.tsx: drop the duplicate SEVERITY_RANK and import the canonical
  one from data/scanners/constants.
- SecuritySafesTable: extract cells (ScoreCell / ThresholdCell / VersionCell
  / BalanceCell), helpers (formatBalance, getEvidence, getAggregateSummary,
  getAggregateSafeGrade, hasMultichainWarning, isAnyChainScanning), and
  visual constants (DASH, ROW_VARIANTS, MotionTableRow, MotionTbody) into
  three sibling files. The main file drops from 684 → 482 lines.
